backend/python-tests/requirements.txt
pytest==8.2.1
requests==2.32.3
python-socketio==5.11.3
websocket-client==1.8.0
